requirements/release.txt
python-dateutil>=2.8.1
flask>=1.1.1
exitstatus>=1.4.0
